MOVIE RATING PREDICTION WITH PYTHON
This project focuses on predicting the ratings of movies based on various features such as genre, director, actors, and other relevant factors. The goal is to develop a regression model using historical movie data to accurately estimate movie ratings as provided by users or critics. By analyzing this data, we can gain insights into the factors that influence movie ratings and build a predictive model.